Add unit tests
* Add unit tests for LPCertificateItemMapper
* Add unit tests for LLPCerticateItemMapper
* Add unit tests for MapUtil
* Refactor unit tests from map.item.service into OtherCertificateItemMapper.unit.test
