+++ src/model/AddressRecordsType.ts
@@ -1,0 +1,6 @@
+export const enum AddressRecordsType {
+    CURRENT = "current",
+    CURRENT_AND_PREVIOUS = "current-and-previous",
+    CURRENT_PREVIOUS_AND_PRIOR = "current-previous-and-prior",
+    ALL = "all"
+}
